requirements.txt
aioch2==0.0.4
aiofiles==0.8.0
clickhouse-driver==0.2.2
fastapi==0.70.1
PyYAML==6.0
uvicorn==0.16.0
